Hack our way around jenkins tmpdirs
The comment basically says it all: the way jenkins uses tmpdirs means
it's not really a good idea for anything to `dir()` to a subdirectory of
webapp while deploy_to_gae.py is running its local changes check.  (I'm
not actually sure if it's the `dir()` or `sh()` inside of it.)  So now
we just have the shell do the cd to avoid the issue.  Ugly, but it
should work.

Test plan: cross fingers, deploy with SERVICES=kotlin-routes,dynamic

Reviewers: csilvers, kai

Subscribers: #devops