Add kotlin-routes to deploy-webapp.groovy
Summary:
The big day approaches!  It's time to allow doing the set-default of
kotlin-routes through the main webapp deploy process.  Since
set_default.py doesn't really do parallel stuff right now, I did the
parallel gradle command in groovy, which is the bulk of this diff.  The
rest is just updating variable names, definitions, and usage to work
better with services.

This is the companion diff to D46049.  (They may be deployed in either
order, as long as SERVICES=kotlin-routes is not set until both are
deployed.)

Test Plan:
- push this, make sure deploys don't see any problems
- deploy D46049 with SERVICES including `kotlin-routes`, monitor
  carefully to make sure things go ok

Reviewers: csilvers, colin

Differential Revision: https://phabricator.khanacademy.org/D46056