Try out the html test report plugin for khanalytics core tests
Summary:

When running tests via gradle, only minimal output is printed to the
console, and additional info is available in an html report, but this
only lives in the workspace on the particular test worker that ran the
tests, so it's pretty useless.

To get around this, I installed the HTML publisher plugin
(https://wiki.jenkins.io/display/JENKINS/HTML+Publisher+Plugin)
which makes these reports visible in the jenkins UI. This commit adds
some code to the khanalytics-test.groovy job to upload the test report
for the `core` project. If this all looks good, I'll set up some build
script code to combine the test reports from all the projects together
and then use that for the report.

Test Plan:
- push and try it!

Auditors: ragini, kelsey, tom