Do a second `git clean` to help clean up submodules with .gitignores.
When we are cleaning out a submodule, *and* that submodule disappeared
as part of this `git checkout -f`, *and* that submodule had a
.gitignored file in it (such as a .pyc file) then the clean above
won't totally delete this submodule because it will leave around the
.pyc file.  BUT it will delete the .gitignore, so a subsequent git
clean -ffd will delete it.  We always just clean twice to be safe.

This scheme doesn't work for files excluded via a global .gitignore,
rather than a local one.  It's still a TODO to come up with a plan for
that as well.

Auditors: benkraft

Test Plan:
./safe_git_test.sh