Do automatic commits in a (new) branch, not directly on master.
Summary:
We were having a problem that buildmaster got upset that new commits
were showing up in master while a deploy was going on.  So our new
rule is that when we do automatic commits, we never do them straight
to master, but always to some other branch -- like we've always done
with translations.  Then we merge that branch in during the next
deploy, again just like with translations.

I cribbed the logic from update-i18n-lite-videos.groovy, one of the
scripts that was already doing this for `translations`.

This is the change to commit to the new branch.  The change to merge
it in automatically will happen in the buildmaster repo.

I manually pushed a new `automatic-commits` branch to github so these
should succeed tomorrow.

Check in translations changes to the automated-commits repo as well.

This will make buildmaster life easier (and deploys less confusing) --
now *all* automated commits go to this new branch, both the ones that
update intl/translations (which used to be in the 'translations'
branch) and the new ones: weekly-mainteance and
determine-test-dependencies.

I *think* this should be safe -- we already update the translations
branch via several jobs that can be running in parallel, so adding a
few more shouldn't hurt.  But if not we can roll back.

Test Plan:
Fingers crossed.  I ran
   groovy jobs/determine-test-dependencies.groovy
   groovy jobs/webapp-maintenance.groovy
but the real test will be to see what determinte-test-dependencies
does after this is landed.

Fingers crossed

Reviewers: benkraft, drosile, amos

Reviewed By: benkraft, drosile, amos

Subscribers: yeva

Differential Revision: https://phabricator.khanacademy.org/D46461