Add JOB_PRIORITY parameter to deploy scripts
Summary:
(See https://phabricator.khanacademy.org/D46578 for more context).

With the jenkins priority sorter plugin, we can now add priorities to our
jobs, so that we can reorder the queue as needed. This just adds a
description of what the JOB_PRIORITY parameter is and its suggested value.

To make sure we run our jobs in the same order as our usual pipeline, we
assign different values to the JOB_PRIORITY depending on the job (if
it is related to a job that is currently deploying). A lower number
indicates a higher priority.

* merge-branches: 1
* webapp-test/build-webapp: 2
* deploy-webapp/e2e-test: 3

Depends on D46578.

Test Plan: pretty much yolo, suggestions for testing welcome!

Reviewers: jacqueline, csilvers

Differential Revision: https://phabricator.khanacademy.org/D46582