Compress the jenkins logs every week.
Summary:
We have a plugin that tries to compress the `log` file, but it doesn't
compress the `#.log` file, and various errors can cause it not to
compress the `log` file either.  So let's just do that via a weekly
cronjob!

Test Plan:
I ran this on jenkins as the jenkins user:
   % cd jobs/misc/jobs/webapp-maintenance/builds
   % find . -mtime +1 -a -type f -a \( -name 'log' -o -name '*.log' \) -print0 | xargs -0rt gzip
and it started compressing a bunch of stuff.

Reviewers: jacqueline, benkraft

Reviewed By: jacqueline

Differential Revision: https://phabricator.khanacademy.org/D46636