Increase the khanalytics deploy timeout to 90 minutes
Summary:

We've added two new containers (R and monitoring) to the build this week, and
now it's timing out. Let's just up the timer as a stopgap. I'm going to see if
I can parallelize any of the build to speed this up a bit more.

Test Plan:
- fingers crossed

Auditors: ragini