Run khanalytics image builds in parallel
Summary:

This is part 2/2 of a commits to try to parallelize the khanalytics build so it
doesn't take as long. In a previous commit in the khanalytics-private repo, I
changed some of the deploy scripts to break up steps, in order to allow us to
coordinate more of the build via jenkins.

This commit changes the khanalytics deploy job to call these broken down steps,
doing the slow part of building each image in parallel.

Test Plan:
- fingers crossed (try deploying; you'll probably see follow up bugfixes)

Auditors: ragini, kelsey