Add JOB_PRIORITY to khanalytics-test job
Summary:
For consistency, we add a JOB_PRIORITY parameter to khanalytics test and deploy
jobs. Sharing is caring, y'know?

Right now the default priority is set to 9, so if any other deploy jobs are in the queue, they'll starve out
our khanalytics jobs and we don't want that.

Test Plan: fingers crossed

Reviewers: benkraft, csilvers, colin

Reviewed By: colin

Subscribers: colin

Differential Revision: https://phabricator.khanacademy.org/D46865