Exclude 'manual only' tests when running dev-e2e tests
Summary:
If we pass `all` as a test type to the webapp-test job, it includes
testing flow and js tests. When running the dev-e2e test job, we don't
want to include these, but we also don't want to just run 'relevant'
tests, so this patch adds a new parameter to let us do that.

Test Plan:
Run on jenkins, make sure that flow and js tests are not part
of the tests run.

Reviewers: csilvers, benkraft

Differential Revision: https://phabricator.khanacademy.org/D46879