Ensure that we use a local variable for the image
Summary:

Apparently loop variables in groovy are not local to the loop, so if you
capture the variable in a closure (as we do for the khanalytics deploy image
name), you end up with all the closures capturing the same variable value.

This failed the deploy as all the parallel stages tried to use the same sandbox.

This commit creates a new local variable outside the closure for each loop
iteration and then captures this in the closure.

I also created a new parameter to skip tests, with a note to use this only for
debugging the deploy system, so that the debugging cycle on this script is not
so painfully slow.

Test Plan:
- fingers crossed (deploy again)

Auditors: ragini, kelsey