Only set test file glob if it differs from the default
Summary:
We run tests on many branches that do not contain master, so do not
currently have the `--test-file-glob` flag that I added. This causes
them to fail and backs up the test worker queue. In this patch we change
the behavior to only set the flag if the desired glob differs from the
default. We should be able to remove this conditional logic once enough
folks merge master into their branches.

Test Plan: Fingers crossed!

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D46913