Make timeout longer for allcheck
2 hours isn't long enough anymore!  Which is sad -- and maybe something
we should try to improve -- but having it crash is the worst of all
possible worlds.  I kicked the can by bumping it to 4 hours.  To be a
more responsible can-kicker, I did this only if we are running huge
tests (make_test_db_test is the only problematic one) and put it back
down to 1 hour otherwise.

Test plan:
- `groovy jobs/webapp-test.groovy` gives only the usual errors
- allcheck should not time out tomorrow

Auditors: csilvers, carter, yeva