Bump the number of e2e-test builds we keep
In the future, we may do something cleverer, or put more data in a
useful place, but this is easy.  Currently the logs are 28GB so this
should add about 42GB, of 220GB currently free on Jenkins, so we won't
have to expand our disk just yet.

Test plan: fingers crossed

Auditors: jacqueline, mita