Run end-to-end tests earlier in deploy
Now that we've set up cookie splitting, we can start end-to-end tests as
soon as we start migrating traffic -- usually about 10 minutes before
we currently do.  This script sets things up to do so.  It's pretty
straightforward: a new script in webapp (in a forthcoming diff) will do
all the interesting parts.  We just have to wait for it.  I also added a
timeout on the wait-script, just in case, although if the version is
actually not changing, there are many more warnings before this one.

Test plan:
- cross fingers
- deploy with this (and the companion diff) and confirm that:
  - the e2e tests start when I expect (just as traffic starts to
    migrate)
  - they run against the new version (looking at the logs)

Reviewers: jacqueline, mita, dhruv