Update the test jobs to use the new "expected tests" option in summary
Summary:
This is the jenkins half of D48011.  In this diff I actually use the
flag created there.  In the case of webapp-test, I left in the old
method as a fallback, since people sometimes run tests on old version.
(I removed a previous fallback for --slack-thread, since it's been a few
months now.)  I also left a similar method in in the case where all
workers crash, although that should be rare.

Test Plan: cross fingers, keep an eye on dev-support-log

Reviewers: #devops

Subscribers: mita

Differential Revision: https://phabricator.khanacademy.org/D48012