Set the traffic-split cookie in the GraphQL integration tests
We need this magic cookie to get the "new" side of the traffic split, if
we run before all traffic is migrated.  (This is unlikely to matter:
usually these tests work exactly the same before and after set-default
so we would see any issues earlier, but it could cause spurious
failures.)  We should also do this for android, but it's a little less
important there since everything should be backwards-compatible so we
shouldn't get spurious failures; and in any case I will need some mobile
help to do so!

Test plan: crossed fingers

Auditors: csilvers