Update the label for e2e tests to include more detail
Summary:
This was requested by QA (a part of INFRA-1817).  We now include whether
this is a pre- or post-deploy test, and the URL we are using.  I just
pass this into the existing "label" attribute which is used by
`tools/test_pickle_util.py summarize-to-slack`.  While I was at it, I
replaced `--commit` with `--label`, its non-deprecated synonym.

Test Plan: cross fingers, watch the results after deploying this

Reviewers: #devops, csilvers

Reviewed By: #devops, csilvers

Differential Revision: https://phabricator.khanacademy.org/D48052