Add Sourcemap Upload and Error Version Flags
Summary:
Adding the sourcemap upload flag to the build-webapp script so that we static code changes, we properly upload the sourcemap files to Sentry

Also adding in the related error version tracking flag. By default we update the error version number everytime we do a GCS upload, but we don't want to do it if we didn't build and upload the sourcemaps in the build-webapp step. So, if we're not deploying static code, we do not update the error reporting versioning

Test Plan: build webapp and set-default

Reviewers: benkraft, csilvers

Subscribers: jacqueline, amos