Update weekly-maintenance to keep any static files updated recently
Summary:
This is to match D48757, and will be deployed around the same time or
earlier.  I updated some comments to make intentional the fact that the
manifest may not include itself, and, more importantly, made sure we
keep any files recently uploaded in case they are part of a running
deploy, since we now definitely upload the manifest after all the other
files.  (I did "yesterday or today" to make things easy to do with grep
and include at least 24 hours.)

Test Plan: cross fingers

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D48764