New job which creates all.pot.pickle in i18n-update-strings/genfiles/translations/new path.
Summary: Dividing i18n-update-strings into multiple jobs (this is a first trial divvy).

Test Plan:
Configure this job to run after the current scheduled i18n-update-strings job, and compare the all.pot.pickle generated.
If both the all.pot files are the same, replace the job with a new one.

Subscribers: tom

Differential Revision: https://phabricator.khanacademy.org/D48781