Modify arg format for notifySetDefault
Summary:
I think groovy actually reads named args as a map, and therefore is maybe not sending the sha as we expect causing buildmaster to 400 from the schema validation.

Fingers crossed this solves the problem with the second round of smoke tests not getting triggered.

Auditors: csilvers, benkraft

Test Plan:
Fingers crossed