Send unstable smoke test builds as status failed
Summary: The way jenkins currently notifies buildmaster is that an e2e-test job that results in a status of "unstable" (successfully completed job but with failed tests) is sent to buildmaster as "success". But this is very misleading as deployers read this as "my tests passed". Instead we should mark those as failed.

Test Plan: Fingers crossed

Reviewers: benkraft, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D48901