Make sure test-workers work at GIT_SHA1, not at master.
Summary:
I was doing something silly before, where we'd sync to GIT_SHA1 and
then immediately sync to origin/automated-commits instead.  We did
this because I was thinking fuzzy when I introduced the
automated-commits branch here.  In reality, we only need to work on
automated-commits when creating and checking in the new file!  That's
what we do now.

Test Plan: Fingers crossed

Reviewers: benkraft

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D48915