Notify the buildmaster when we start setting default
Summary:
This is basically an un-revert of c289c62987, except that several bits
have been modified significantly to match the new way we do e2e tests.
It depends on D49024.  I left the old notification in, for reasons
described in the other diff.

I also added back the parameters to the e2e-test job which are needed
for the new path, reverted in that same commit.  (I likely need not have
reverted them there.)

Test Plan:
cross fingers, deploy it, check the buildmaster has the right status at
various times during set-default.

Reviewers: jacqueline

Reviewed By: jacqueline

Differential Revision: https://phabricator.khanacademy.org/D49030