New job which creates all.pot.pickle in (i18n-update-strings/genfiles/translations/new) path.
Summary:
In order to debug and optimize the i18n-update-strings job, I am dividing i18n-update-strings into multiple jobs.
The reasons to do so:
1) Lesser timeouts : Currently the job times out and corrupts the binary files in the process, Dividing into smaller jobs should help with this.
2) Easier to rerun only the failed jobs : Currently if the i18n-gcs-upload(last step) fails, we still need to rerun the entire job, which is a waste of time and resources. The goal is to have only the failed jobs rerun
3) Easier to debug and fix issues: Currently the team has spent considerable time and effort in figuring out issues, which is not easy as it is difficult to isolate the causes. The smaller jobs with better debugging messages should help with saving eng time.
4) Removing Dropbox dependencies and creating a single `source of truth`

Steps taken:
This is the first step of the job.
It should only create the all.pot.pickle file in a new path (for now).
I am hoping to divide this job later into 2 as well (one that creates the all.pot file and another that syncs the graphie images).
Once this is successful, I will update the i18n-update-strings job to use this new file and also divide the other jobs.

Test Plan:
Please take note that https://phabricator.khanacademy.org/D48782 has to be deployed in webapp before this changes are run.
Configure i18n-create-all-pot-pickle to run after the current scheduled i18n-update-strings job, and compare the all.pot.pickle generated.
If both the all.pot files are the same, we will update the new jobs.

Reviewers: michaelpolyak, csilvers

Reviewed By: csilvers

Subscribers: csilvers, tom

Differential Revision: https://phabricator.khanacademy.org/D48781