Remove end slash from e2e URL param if included
Summary: E2e jobs are often run manually from the jenkins UI, which means URLs are not entered uniformly, and are sometimes entered with a slash at the end, which is not what the e2e-test.groovy script expects. If that is included, remove it from the URL param.

Test Plan: Try rerunning https://jenkins.khanacademy.org/job/deploy/job/e2e-test/9385/

Reviewers: benkraft, csilvers

Reviewed By: benkraft, csilvers

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D49124