Deploying kotlin to appengine standard, part 2: jenkins deploy/set-default
Summary:
In D49391, I started setting up changes to set-default and priming so that we
can deploy kotlin to webapp.  This commit is a corresponding change to
jenkins-jobs that moves the deployment and set-default over to the new system.
This ended up being a small simplification, since there's less special casing
of 'static' in the groovy code now.

Still TODO: rollback in both jenkins-jobs and in webapp.

Test Plan:
- fingers crossed? I'll see how it goes when I deploy the new process.
- (testing ideas very welcome!)

Reviewers: benkraft, csilvers

Reviewed By: benkraft, csilvers

Differential Revision: https://phabricator.khanacademy.org/D49441