Fix pingForStatus so jenkins can find out state of smoke tests
Summary:
Upon launching the integration of smoke tests with buildmaster, pingForStatus was failing with an error message along the lines of `don't use getAt`. But turns out hidden in the docs for the Jenkins httpRequest plugin are a couple of nice helper methods for handling responses, including getStatus and getContent (and presumably didn't like that I was using indexing to access response data).

With this fixed we can turn the verifySmokeTestResults step back on so that deploy jobs wait for smoke tests to complete before prompting deployers for `sun: set default` and `sun: finish up`.

Test Plan: Fingers crossed

Reviewers: benkraft, csilvers

Reviewed By: benkraft, csilvers

Differential Revision: https://phabricator.khanacademy.org/D49543