Fix bug in waiting for second-smoke-tests to finish
Summary:
Waiting for first smoke tests to finish before asking a deployer to set default was working properly while waiting for second smoke tests for finish up is not. Jenkins logs show that it was checking the status of the first smoke tests when it should have been looking at the second, so something must be amiss in how we're passing the setDefault value to verifySmokeTestResults.

But that's ok because its actually more straightforward to just explicitly pass it the jobName instead of a setDefault flag.

Test Plan: fingers crossed

Reviewers: csilvers, benkraft

Reviewed By: csilvers, benkraft

Differential Revision: https://phabricator.khanacademy.org/D49647