Retry requests to the buildmaster from Jenkins
The buildmaster sometimes flakes.  Now we'll retry!  For now, I just
retry everything, because it was too hard to figure out how to detect
whether the error is 4xx in a clean way.  Most of those 4xxes are pretty
fast, anyway, so it's not a big cost.

Test plan: cross fingers

Reviewers: csilvers, jacqueline, kai