Retry requests to the buildmaster from Jenkins
Summary:
The buildmaster sometimes flakes.  Now we'll retry!  For now, I just
retry everything, because it was too hard to figure out how to detect
whether the error is 4xx in a clean way.  Most of those 4xxes are pretty
fast, anyway, so it's not a big cost.

Test Plan: cross fingers

Reviewers: csilvers, jacqueline, Kai

Reviewed By: csilvers, jacqueline, Kai

Differential Revision: https://phabricator.khanacademy.org/D50006