Add -- and use -- a `--verbose` option to timeout_output.py.
I think this timeout is triggering on our nightly allcheck builds, but
I can't be sure because we're not seeing the return-value of the
timeout_output call, and there's no message printed on failure.

I decided to solve the problem by addressing the latter failing.  Now
we print a message on failure, but only in (a new) verbose mode, which
we now use for all our tests.

Auditors: colin, benkraft

Test Plan:
I ran
   time ./timeout_output.py 1s sleep 3; echo $?
and it failed after 1 second and gave an rc of 124.  This is unchanged
from before this commit.

Then I ran
   time ./timeout_output.py -v 1s sleep 3; echo $?
and it did the same thing, but also emitted
   KILLING process 5681 after 1.0 seconds of no output to stdout