Have all test workers clone the mobile repo
This is a companion to D50424, where I convert the android integration
test to a common ordinary integration test.  In this commit, we simply
clone/update the mobile repo on each test-worker, in case it has to run
the android test.  I'll deploy this before D50424.  In a separate
commit, I'll remove the old version of the test, after the new one has
been working for a bit.

Test plan: cross fingers, deploy D50424

Reviewers: csilvers, jacqueline

Subscribers: hannah, jared, dhruv, mita