Add a dry-run mode to the emergency rollback job
The primary purpose of this is to make sure the repos are up to date,
which can take quite a while otherwise.  I intend to run it every week
on a cron job (from buildmaster, since I don't think we can use
parameters in jenkins cron) to keep things warm.  (I'll add that in a
separate diff.)

Test plan:
- `groovy jobs/emergency-rollback.groovy`
- run emergency rollback with DRY_RUN, should do the clone and stuff
- run emergency rollback for reals (will batch with D50426), should be
  faster

Reviewers: csilvers, amos