Always call rollback.py when aborting deploy
Summary:
Up to now, when a deployer would do an abort during a deploy-webapp job, we had Jenkins call the current_version.py script which would return which version is live to help us determine if it is necessary to run a rollback.

This was causing a major issue in neglecting necessary rollbacks, as it wasn't accurate detecting when our traffic was split between two versions. When say, 17% of traffic had been migrated to the new version, but not the rest, the call to the KA /_api/version endpoint would often hit one of the instances from the other 83% and therefore erroneously tell us we're still entirely on the old version and make the faulty conclusion that we don't need to run a rollback.

In this commit, we scrap the check on current version in favor of always requiring our deploy-webapp to trigger the deploy/rollback.py script. And then if it doesn't really need to rollback traffic, that script will just exit early. The exception to this change is when an abort is triggered before set-default. We now catch that separately and finish without rollback.

Test Plan:
This change will not be deployed until `deploy/rollback.py` has been run manually to verify script appropriately handles case where there is no traffic split and therefore no rollback required.

Step to do this:
1. Build/test a branch as normal
2. Wait on setting default (so that there is guaranteed 0% traffic on the new version)
3. Simulate what it would be like if we had set default but aborted before any traffic started migrating, using something along the lines of:

`deploy/rollback.py --bad 'gae-181210-0930-fb574926318f-kotlin-routes-181205-1526-f5bb954b393b' --good 'gae-181210-0753-7ac172ba783a-kotlin-routes-181205-1526-f5bb954b393b'`

Note: The tags themselves will of course be different, they will be the new version (bad) + the version that still has traffic on it (good).

4. Verify that the following is logged:

  - "Tagging <new (bad) version> as bad in git"
  - "Rolling back from <bad version> to <good version>"
  - "Module default already set to <good version>, exiting."

Reviewers: benkraft, csilvers

Reviewed By: benkraft

Differential Revision: https://phabricator.khanacademy.org/D50911