Accept version params in emergency rollbacks
Summary:
Our emergency rollback jenkins job, to date, has only allowed us to rollback to one version prior. If Khan engineers found a problem that was introduced in an earlier deploy, we had no good way other than running the deploy/rollback.py script manually to allow them to select an earlier version to rollback to.

This commit allows you to specify which version is the good version to rollback to, and which version should be marked bad.

Test Plan:
This change should not affect the way the emergency-rollback job works currently (with no params provided). It should be safe to deploy and test first with DRY_RUN=true to ensure that the deploy/rollback.py script is receiving the expected version args.

To be safe, this will still be deployed and tested outside of regular deployment hours.

The following will be tested in DRY_RUN mode:
a) Both ROLLBACK_TO and BAD_VERSION are provided and are valid tags, which should result in rollback.py being called with both `--good` and `--bad` args.
b) Just one of the two params are provided, such as ROLLBACK_TO, which should result in rollback.py being called with just the `--good` arg
c) Both provided, but BAD_VERSION does not specify a dynamic version, for example `static-181207-1015-1a1946c285df`. This should fail before running the rollback.py cmd.
d) ROLLBACK_TO defines a gae version that does not exist, for example an outdated one like:
`gae-181206-1106-de425f073eaf-static-181206-1113-863938f49c2e-kotlin-routes-181205-1526-f5bb954b393b`.
This should also fail before running rollback.py.
e) Neither params provided, which should result in rollback.py to be called but with no args.

Reviewers: benkraft, csilvers

Reviewed By: benkraft, csilvers

Subscribers: amos

Differential Revision: https://phabricator.khanacademy.org/D50922