Allow saying you want to wait longer
This is a stop-gap: we have Grand Plans to make the waiting process more
interactive.  But for now, we want to be able to have more time for
known-risky deploys.  So you can just say `WAIT_LONGER=true` and it will
give you 6 hours.

Test plan: cross fingers, `groovy jobs/deploy-webapp.groovy`

Reviewers: jacqueline, csilvers, amos, boris, andrew