Allow running smoke tests against a dev server
Summary:
This is a companion to D51220: it hooks the flag up to a jenkins option.
For now, this will just be used manually -- the main thing is to get it
passing and then we can maybe run it more regularly.

Test Plan: run the jenkins job, checking the "DEV_SERVER" box in the UI

Reviewers: csilvers, jacqueline, mita, dhruv

Reviewed By: csilvers, jacqueline, mita

Differential Revision: https://phabricator.khanacademy.org/D51221