Remove set-default prompt from deploy-webapp
Summary:
See corresponding buildmaster changes at: https://phabricator.khanacademy.org/D51575

In an effort to have much of the Slack communication owned by buildmaster and not split between buildmaster and jenkins, this is the first in a larger change of shifting set default and finish up prompt handling over to buildmaster.

This change includes removing the set-default prompt originally sent from jenkins. It also revises how jenkins checks that a deployer has completed manually testing by pinging buildmaster fot the set_default_prompt_status. Similar to how it checks for smoke test results, it will check periodically (in this case once every 10s) until it finds a `confirmed` status for the prompt (or in the case of buildmaster outage, it will trigger a manual confirmation step to allow a deployer to proceed thru the Jenkin's UI). So that the changelog does not arrive after a deployer is prompted to do manual testing, we also move send changelog to the build job.

We are leaving the finish up prompt in place until these changes are deployed and any kinks are worked out. Therefore, this commit does not remove much of the scaffolding around how jenkins monitors timing.

Test Plan: Fingers crossed

Reviewers: benkraft, csilvers

Reviewed By: benkraft, csilvers

Differential Revision: https://phabricator.khanacademy.org/D51576