Increase some wait-times in build_current_sqlite
Summary:
We're having a problem with the new datastore emulator where sometimes
the dev server is unresponsive (or specifically the datastore emulator
is unresponsive) when we try to sync spanish.  I don't really know
what's up, but I'm hoping that we just need to wait a bit longer -- we
kill the one dev server and immediately start the next one.  So, now, I
bump the wait times.  The job already takes hours anyway so this
shouldn't be a big deal.

Test Plan: run the job, hope it syncs spanish

Reviewers: tom, csilvers

Reviewed By: csilvers

Differential Revision: https://phabricator.khanacademy.org/D52530