Stop sending finish up prompt from jenkins
Summary:
Addresses Jira ticket: INFRA-2590

In D52713, we shift the responsibility of sending out the finish up prompt from jenkins to buildmaster. This commit removes this step from the jenkins deploy-webapp job. It also adds a notify buildmaster call to send an update when monitoring is complete, as this notification is required for buildmaster to trigger the prompt.

Note: There is a followup task to remove unneeded slack messages + code after the prompt shift is complete (see INFRA-2679).

Test Plan: Fingers crossed

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D52761