Make it obvious when jenkins cannot reach buildmaster
Summary: If Jenkins pings buildmaster and gets consecutive failures (presumably due to a buildmaster outage), it typically falls back to allowing a deployer to proceed in the Jenkins logs. This, however, is not made apparent in Slack and means the deployer might not realize they need to do it through Jenkins until they've been waiting for a significant amount of time. This commit has Jenkins ping Slack directly in the case of a buildmaster outage.

Test Plan: fingers crossed

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D53618