Split BASE_REVISION into two params
Summary:
Addresses Jira ticket: INFRA-2767
Corresponding buildmaster diff: D53747

This bug fix was prompted by a longstanding issue arising during build-webapp jobs for any dynamic build following a tools-only deploy. With a dynamic-only build job, we attempt to find the proper static version for --copy-from. But if our parent commit was tools-only, we end up referencing a static version that doesn't exist causing our build to fail. It looks for a manifest with the tools-only version name, which of course doesn't exist.

One root of this problem is that our build-webapp job was trying to do two different things with BASE_REVISION: 1) Use it to determine all changes since the previous revision for the changelog and should_deploy, and 2) Use it to copy static or dynamic version names when we aren't deploying a particular service (i.e. copy a static version name when not deploying to static).

The fix introduced in this commit is to pass a new COPY_FROM_REVISION param, so as to ensure we'll have the correct version name for a valid static and dynamic build, and pare down the responsibility of BASE_REVISION, using it now only for the first use case and not the second.

Test Plan: Fingers crossed

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D53748