an wrapper to call the buildmaster from Jenkins
Notes: I did not test it yet, just want to send the diff early
for review to make sure it makes sense to you or not.
If the buildmaster is down, we need to ping dev-support loudly.

Summary:
to have a wrapper to call _makeHttpRequest to the buildmaster.
If the buildmaster is down, alert loudly and ping @dev-support
in "#infrastructure-devops" slack channel.

Test Plan: to test buildmaster.groovy in Jenkis console or in a ZND job.

Reviewers: jacqueline, benkraft

Differential Revision: https://phabricator.khanacademy.org/D53692