Work around a nasty bug causing deploy rollback to not execute
The jenkins bug is described more succinctly in
https://issues.jenkins-ci.org/browse/JENKINS-57163, but basically, some
interaction of timestamps, parallel, sleep, and exception handling in
groovy causes only part of our `catch` block in `deploy-webapp` to
execute.  The effect of this is to not do the rollback on abort, which
is pretty bad!  For now, we work around it by removing the timestamps on
the background thread, which in my simple test case avoided the bug.
Hopefully it will work here.

Test plan: fingers crossed (plus as described in the upstream bug)

Auditors: jacqueline, kai, csilvers