Have tools-only deploys verify smoke tests and set-default prompt
Summary:
As it currently stands, tools-only deploys do not wait and verify if your smoke tests are complete or even if you told sun to set default (though we do run first_smoke_tests and you do receive a set default prompt).

Let's change that to ask the deploy to verify both the smoke tests and that the prompt has been confirmed.

Test Plan: Fingers crossed

Reviewers: benkraft

Reviewed By: benkraft

Subscribers: csilvers

Differential Revision: https://phabricator.khanacademy.org/D53774