Run generate topic tree json for all locales driven by the groovy pipeline script.
This change is necessary because the prior version where the generate topictree json call from python by iterating over all
locales that are TEST or better in devshell fails due to http error. It looks like devshell gets in a state where after
a large number of rpcs it starts failing. This change is to invoke devshell process one locale at a time from the wrapper
script.

Auditors:
neha, aasmund