Fix the git ops in update-ownership-data
We were doing the git ops in the wrong order (I had cribbed poorly from
determine-test-dependencies) which meant we were clobbering our changes
just before trying to publish them.  Now we don't: we do our checkouts
earlier and avoid the problem.  (`safeCommitAndPush` still does a pull
in case there were other changes committed in the meantime, so it's not
a problem to do this earlier.)

Test plan: run the job; should see a commit on `automated-commits`

Auditors: amos, csilvers