Use `gsutil -m` to delete faster.
webapp-maintenance timed out last week, and it seems to be because
clean-ka-static took 8 1/2 hours:
   https://jenkins.khanacademy.org/job/misc/job/webapp-maintenance/35/flowGraphTable/

This may because there's a lot of extra files left around in the new
webpack build.  In any case, the logs said:
```
03:19:53  ==> NOTE: You are performing a sequence of gsutil operations that may
03:19:53  run significantly faster if you instead use gsutil -m rm ... Please
03:19:53  see the -m section under "gsutil help options" for further information
03:19:53  about when gsutil -m can be advantageous.
```
so let's try that!  I don't see any reason not to do it; I think it
was just an oversight bfeore.

Auditors: benkraft, kai

Test plan:
Fingers crossed