Conditionally run the LTT Update job as part of translation-analytics.
Summary:
The LTT Update job finds revisions in the locale's stage that have
fully-translated updates in English and imports them.

The LTT Update job has been running in webapp, however it needs to be fairly
closely integrated with TAP runs because it A) depends on the EN LTT TAP data,
and B) assumes that the locale stage TAP data is updated immediately following
the LTT update. The best way to ensure this happens in the right order is to
just always run the update job in sequence with the TAP jobs. This has the
upside that when a revision becomes fully translated (due to new translations
being downloaded from Crowdin) the update happens immediately as
translation-analytics is triggered as part of the download.

LTT updates are still opt-in, based on a list that was moved into here from
webapp.

Test Plan:
Crossed fingers.

Reviewers: aasmund

Subscribers: